fix uint16 type define.

+++ serial.go
@@ -20,7 +20,7 @@ var (
 	int8Type    = reflect.TypeOf(int8(0))
 	uint8Type   = reflect.TypeOf(uint8(0))
 	int16Type   = reflect.TypeOf(int16(0))
-	uint16Type  = reflect.TypeOf(uint64(0))
+	uint16Type  = reflect.TypeOf(uint16(0))
 	int32Type   = reflect.TypeOf(int32(0))
 	uint32Type  = reflect.TypeOf(uint32(0))
 	int64Type   = reflect.TypeOf(int64(0))
